pyproject.toml
[project]
name = "core"
version = "0.1.0"
description = "Core-Platform with User and Location"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.13",
    "celery>=5.5.3",
    "cerberus>=1.3.7",
    "cryptography>=45.0.5",
    "django>=5.2.4",
    "django-cors-headers>=4.7.0",
    "django-db-connection-pool>=1.2.6",
    "django-environ>=0.12.0",
    "django-extensions>=4.1",
    "django-guardian>=3.0.3",
    "django-health-check>=3.20.0",
    "django-ratelimit>=4.1.0",
    "django-redis>=6.0.0",
    "djangorestframework>=3.16.0",
    "drf-spectacular>=0.28.0",
    "drf-spectacular-sidecar>=2025.7.1",
    "flower>=2.0.1",
    "hiredis>=3.2.1",
    "httpx>=0.28.1",
    "kombu>=5.5.4",
    "marshmallow>=4.0.0",
    "phonenumbers>=9.0.8",
    "pillow>=11.3.0",
    "prometheus-client>=0.22.1",
    "psycopg2-binary>=2.9.10",
    "pyjwt>=2.10.1",
    "python-decouple>=3.8",
    "python-dotenv>=1.1.1",
    "pytz>=2025.2",
    "redis>=6.2.0",
    "sentry-sdk>=2.32.0",
    "structlog>=25.4.0",
]
